Adding local storage support

+++ icon_one.php
@@ -12,22 +12,25 @@ if ($image) {
   $json_mochiswap = file_get_contents('https://s3-us-west-1.amazonaws.com/tokens.mochiswap.io/hashparty-default.tokenlist.json');
   $infoMochiswapData = json_decode($json_mochiswap);
   foreach ($infoMochiswapData->tokens as $infoMochiswap) {
-    if ($selectedToken === strtolower($infoMochiswap->address)) {
+    if (strtolower($selectedToken) === strtolower($infoMochiswap->address)) {
       $image = file_get_contents($infoMochiswap->logoURI);
     }
   }
-}
-if ($image) {
-  $file = 'icons/harmony/' . $selectedToken . '.png';
-  file_put_contents($file, $image, FILE_APPEND | LOCK_EX);
-  $file = 'icons/harmony/' . strtolower($selectedToken) . '.png';
-  file_put_contents($file, $image, FILE_APPEND | LOCK_EX);
-} else {
-  if ($autoResolve === 'false') {
-    http_response_code(404);
-    die();
+
+  if ($image) {
+    if (file_exists('icons/harmony/' . $selectedToken . '.png')) {}else{
+      $file = 'icons/harmony/' . $selectedToken . '.png';
+      file_put_contents($file, $image, FILE_APPEND | LOCK_EX);
+      $file = 'icons/harmony/' . strtolower($selectedToken) . '.png';
+      file_put_contents($file, $image, FILE_APPEND | LOCK_EX);
+    }
   } else {
-    $image = file_get_contents('icons/unknown.png');
+    if ($autoResolve === 'false') {
+      http_response_code(404);
+      die();
+    } else {
+      $image = file_get_contents('icons/unknown.png');
+    }
   }
 }
